Add proper logging
