+++ Makefile
@@ -2,11 +2,11 @@ COMMIT=$(shell git rev-parse --short HEAD)
 
 clean:
 	rm -rf **/target
-	rm docker/server/geotrellis-server.jar
+	rm docker/server/geotrellis-server-http4s.jar
 
 build:
 	sbt assembly
-	ln -f server/target/scala-2.11/geotrellis-server.jar docker/server/geotrellis-server.jar
+	ln -f http4s/target/scala-2.11/geotrellis-server-http4s.jar docker/server/geotrellis-server-http4s.jar
 	docker-compose build
 
 serve: build
