+++ build.sbt
@@ -24,20 +24,29 @@ lazy val commonSettings = Seq(
   ),
   resolvers ++= Seq(
     Resolver.bintrayRepo("bkirwi", "maven"), // Required for `decline` dependency
+    Resolver.bintrayRepo("azavea", "maven"),
+    Resolver.sonatypeRepo("releases"),
     "locationtech-releases" at "https://repo.locationtech.org/content/groups/releases",
     "locationtech-snapshots" at "https://repo.locationtech.org/content/groups/snapshots"
   ),
   addCompilerPlugin("org.spire-math" % "kind-projector" % "0.9.4" cross CrossVersion.binary),
   addCompilerPlugin("org.scalamacros" % "paradise" % "2.1.0" cross CrossVersion.full),
   shellPrompt := { s => Project.extract(s).currentProject.id + " > " },
-  fork := true
+  fork := true,
+  assemblyMergeStrategy in assembly := {
+    case "reference.conf" => MergeStrategy.concat
+    case "application.conf" => MergeStrategy.concat
+    case n if n.endsWith(".SF") || n.endsWith(".RSA") || n.endsWith(".DSA") => MergeStrategy.discard
+    case "META-INF/MANIFEST.MF" => MergeStrategy.discard
+    case _ => MergeStrategy.first
+  }
 )
 
 lazy val root =
   Project("root", file("."))
     .aggregate(
       core,
-      server
+      http4s
     )
     .settings(commonSettings: _*)
 
@@ -45,16 +54,32 @@ lazy val core =
   Project(id = "core", base = file("core"))
     .settings(commonSettings: _*)
     .settings(
-      name := "geotrellis-server-core"
+      name := "geotrellis-server-core",
+      assemblyJarName in assembly := "geotrellis-server-core.jar",
+      libraryDependencies ++= Seq(
+        circeCore,
+        circeGeneric,
+        circeParser,
+        geotrellisS3,
+        geotrellisSpark,
+        cats,
+        catsEffect,
+        mamlJvm,
+        kindProjector,
+        simulacrum,
+        typesafeLogging,
+        logbackClassic,
+        concHashMap
+      )
     )
 
-lazy val server =
-  Project(id = "server", base = file("server"))
+lazy val http4s =
+  Project(id = "http4s", base = file("http4s"))
     .settings(commonSettings: _*)
     .dependsOn(core)
     .settings(
-      name := "geotrellis-server",
-      assemblyJarName in assembly := "geotrellis-server.jar",
+      name := "geotrellis-server-http4s",
+      assemblyJarName in assembly := "geotrellis-server-http4s.jar",
       libraryDependencies ++= Seq(
         http4sDsl,
         http4sBlaze,
@@ -62,9 +87,16 @@ lazy val server =
         http4sXml,
         scalaXml,
         geotrellisS3,
+        geotrellisSpark,
         decline,
         cats,
         commonsIO,
+        concHashMap,
+        kindProjector,
+        pureConfig,
+        typesafeLogging,
+        logbackClassic,
         scalatest
       )
     )
+
