+++ core/src/main/scala/geotrellis/server/core/hack/Hack.scala
