+++ core/src/main/scala/geotrellis/server/core/util/CacheRangeReader.scala
@@ -1,0 +1,22 @@
+package geotrellis.server.core.util
+
+import geotrellis.util._
+
+case class CacheRangeReader(rr: RangeReader, cachedBytes: Array[Byte])
+    extends RangeReader {
+  def totalLength: Long = rr.totalLength
+
+  override def readRange(start: Long, length: Int): Array[Byte] = {
+    val end = length + start
+    if (end <= cachedBytes.length)
+      java.util.Arrays.copyOfRange(cachedBytes, start.toInt, end.toInt)
+    else
+      rr.readRange(start, length)
+  }
+
+  protected def readClippedRange(start: Long, length: Int): Array[Byte] = ???
+
+  override def readAll(): Array[Byte] = {
+    rr.readAll()
+  }
+}
