+++ docker-compose.yml
@@ -3,12 +3,12 @@ services:
   server:
     image: quay.io/geotrellis/server:latest
     ports:
-      - "9876:9876"
+      - "9000:9000"
     volumes:
       - ~/.aws:/root/.aws
     environment:
       TEST_VAR: "env_var"
-      HTTP_PORT: 9876
+      HTTP_PORT: 9000
       HTTP_INTERFACE: "0.0.0.0"
     build:
       context: docker/server
