+++ http4s/src/main/resources/application.conf
@@ -1,0 +1,10 @@
+http {
+    "interface": "0.0.0.0"
+    "interface": ${?HTTP_INTERFACE}
+    "port": 9000
+    "port": ${?HTTP_PORT}
+}
+
+catalog {
+    "uri": "s3://azavea-datahub/catalog"
+}
