+++ http4s/src/main/scala/geotrellis/server/http4s/hack/Hack.scala
@@ -1,0 +1,8 @@
+package org.apache.spark
+
+
+package object rdd {
+  class RDD {}
+}
+
+class Partitioner {}
