+++ project/plugins.sbt
@@ -5,3 +5,6 @@ addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.9.0")
 addCompilerPlugin("org.scalamacros" % "paradise" % "2.1.0" cross CrossVersion.full)
 
 addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.2.27")
+
+addCompilerPlugin("org.spire-math" %% "kind-projector" % "0.9.7")
+
