--- server/src/main/resources/application.conf
@@ -1,10 +1,0 @@
-http {
-    "interface": "localhost"
-    "interface": ${?HTTP_INTERFACE}
-    "port": 9876
-    "port": ${?HTTP_PORT}
-}
-
-catalog {
-    "uri": "s3://azavea-datahub/catalog"
-}
