--- server/src/main/resources/logback.xml
@@ -1,13 +1,0 @@
-<configuration>
-    <appender name="stdout" class="ch.qos.logback.core.ConsoleAppender">
-        <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
-        </encoder>
-    </appender>
-
-    <logger name="geotrellis.server" level="${RF_LOG_LEVEL:-INFO}"/>
-
-    <root level="INFO">
-        <appender-ref ref="stdout"/>
-    </root>
-</configuration>
