--- server/src/main/scala/geotrellis/server/wcs/Constants.scala
@@ -1,10 +1,0 @@
-package geotrellis.server.wcs
-
-object Constants {
-  // Map of incoming format string to normalized format string for supported formats.
-  val SUPPORTED_FORMATS =
-    Map(
-      "geotiff" -> "geotiff",
-      "geotif" -> "geotiff"
-    )
-}
